+++ torch/_inductor/runtime/__init__.py
@@ -1,4 +1,5 @@
 from . import triton_helpers, triton_heuristics
+from .hints import instance_descriptor, ReductionHint, TileHint
 from .triton_helpers import libdevice, math as tl_math
 from .triton_heuristics import AutotuneHint
 
@@ -8,4 +9,7 @@ __all__ = [
     "libdevice",
     "tl_math",
     "AutotuneHint",
+    "TileHint",
+    "ReductionHint",
+    "instance_descriptor",
 ]
